+++ changelog.md
@@ -7,6 +7,8 @@
 新增commons-mybatis-generator模块：
  1. 增加`JavaTypeResolverSupport`类：当我们使用mybatis-generator插件自动生成代码时，`JavaTypeResolverDefaultImpl`类是将数据库表中的`TINYINT`类型默认转换为Java的`Byte`类型。改用`JavaTypeResolverSupport`可支持将数据库表中`TINYINT`类型转换为Java的`Integer`类型
  2. 增加模板文件：`src/main/resources/template`目录增加`mybatis-generator.xml`、`mybatis-generator-comment.ftl`两个模板文件。使用方法参考`src/main/resources/template/READMD.md`文档
+ 3. 新增SelectByParamsPlugin插件：该插件支持selectTotalByParams、selectListByParams、selectPagerByParams三类查询接口
+ 4. 新增BatchInsertPlugin插件：该插件为`com.itfsw.mybatis.generator.plugins.BatchInsertPlugin`插件的精简版，去掉了batchInsertSelective接口。
 
 
 ### SpringBoot/V2.0.2